+++ packages/common/web/src/Base/createBaseComponent.jsx
@@ -28,7 +28,10 @@ const createBaseComponent = ({
   relValue,
   t,
   isMobile,
+  pluginDecorationProps = () => ({}),
+  componentWillReceiveDecorationProps = () => {},
   getEditorBounds,
+  onOverlayClick,
 }) => {
   class WrappedComponent extends Component {
     static displayName = createHocName('BaseComponent', PluginComponent);
@@ -40,6 +43,7 @@ const createBaseComponent = ({
     }
 
     componentWillReceiveProps(nextProps) {
+      componentWillReceiveDecorationProps(this.props, nextProps, this.updateComponentConfig);
       this.setState(this.stateFromProps(nextProps));
     }
 
@@ -166,6 +170,15 @@ const createBaseComponent = ({
       return this.isMe() && !this.duringUpdate;
     }
 
+    handleClick = e => {
+      if (onOverlayClick) {
+        const { componentData } = this.state;
+        onOverlayClick({ e, pubsub, componentData });
+      }
+      const { onClick } = this.props;
+      onClick && onClick(e);
+    };
+
     updateComponentConfig = newConfig => {
       pubsub.update('componentData', { config: newConfig });
     };
@@ -215,8 +228,12 @@ const createBaseComponent = ({
     }
 
     render = () => {
-      const { blockProps, className, onClick, selection } = this.props;
+      const { blockProps, className, selection } = this.props;
       const { componentData, readOnly } = this.state;
+      const { containerClassName, ...decorationProps } = pluginDecorationProps(
+        this.props,
+        componentData
+      );
       const { link, width: currentWidth, height: currentHeight } = componentData.config || {};
       const { width: initialWidth, height: initialHeight } = settings || {};
       const isEditorFocused = selection.getHasFocus();
@@ -224,9 +241,9 @@ const createBaseComponent = ({
       const isActive = isFocused && isEditorFocused && !readOnly;
 
       const classNameStrategies = [
-        PluginComponent.WrappedComponent.alignmentClassName || alignmentClassName,
-        PluginComponent.WrappedComponent.sizeClassName || sizeClassName,
-        PluginComponent.WrappedComponent.textWrapClassName || textWrapClassName,
+        PluginComponent.alignmentClassName || alignmentClassName,
+        PluginComponent.sizeClassName || sizeClassName,
+        PluginComponent.textWrapClassName || textWrapClassName,
       ].map(strategy => strategy(this.state.componentData, theme, this.styles, isMobile));
 
       const ContainerClassNames = classNames(
@@ -237,6 +254,7 @@ const createBaseComponent = ({
           [theme.pluginContainer]: !readOnly,
           [theme.pluginContainerReadOnly]: readOnly,
           [theme.pluginContainerMobile]: isMobile,
+          [containerClassName]: !!containerClassName,
         },
         classNameStrategies,
         className || '',
@@ -286,9 +304,15 @@ const createBaseComponent = ({
             .toLowerCase()
             .indexOf('image') !== -1,
       });
+
       /* eslint-disable jsx-a11y/anchor-has-content */
       return (
-        <div style={sizeStyles} className={ContainerClassNames}>
+        <div
+          style={sizeStyles}
+          className={ContainerClassNames}
+          data-focus={isActive}
+          {...decorationProps}
+        >
           {!isNil(link) ? (
             <div>
               {component}
@@ -301,7 +325,7 @@ const createBaseComponent = ({
             <div
               role="none"
               data-hook={'componentOverlay'}
-              onClick={onClick}
+              onClick={this.handleClick}
               className={overlayClassNames}
             />
           )}