+++ packages/common/web/src/Base/createBasePlugin.jsx
@@ -60,8 +60,7 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
   const settings = { ...DEFAULT_SETTINGS, ...config.settings };
   const helpers = config.helpers || {};
   const isMobile = config.isMobile || false;
-  const { t, anchorTarget, relValue, customStyleFn, getEditorBounds } = config;
-
+  const { t, anchorTarget, relValue, customStyleFn, getEditorBounds, onOverlayClick } = config;
   const toolbarTheme = { ...getToolbarTheme(config.theme, 'plugin'), ...config.theme };
   const Toolbar =
     config.toolbar &&
@@ -100,15 +99,17 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
         isMobile,
       }),
     }));
-  const PluginComponent =
-    config.component && config.decorator ? config.decorator(config.component) : config.component;
+  const PluginComponent = config.component;
 
-  const CompWithBase =
+  const BaseComponent =
     PluginComponent &&
     createBaseComponent({
       PluginComponent,
       theme: config.theme,
       type: config.type,
+      pluginDecorationProps: config.pluginDecorationProps,
+      componentWillReceiveDecorationProps: config.componentWillReceiveDecorationProps,
+      onOverlayClick,
       pubsub,
       settings,
       helpers,
@@ -118,6 +119,9 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
       isMobile,
       getEditorBounds,
     });
+
+  const DecoratedCompWithBase =
+    BaseComponent && config.decorator ? config.decorator(BaseComponent) : BaseComponent;
 
   const InlineModals = config.inlineModals;
 
@@ -134,7 +138,7 @@ const createBasePlugin = (config = {}, underlyingPlugin) => {
         const pluginTypes = [config.type, config.legacyType];
         if (includes(pluginTypes, type)) {
           return {
-            component: CompWithBase,
+            component: DecoratedCompWithBase,
             editable: false,
             props: {
               getData: getData(contentBlock, { getEditorState }),