+++ packages/editor/web/src/RichContentEditor/createPlugins.jsx
@@ -1,6 +1,7 @@
 import { composeDecorators } from 'draft-js-plugins-editor';
 import createFocusPlugin from 'draft-js-focus-plugin';
-import createBlockDndPlugin from 'draft-js-drag-n-drop-plugin';
+import createResizeDecoration from './Decorators/Resize';
+// import createBlockDndPlugin from 'draft-js-drag-n-drop-plugin';
 
 const createPlugins = ({
   plugins,
@@ -16,9 +17,14 @@ const createPlugins = ({
   getEditorBounds,
 }) => {
   const focusPlugin = createFocusPlugin();
-  const dndPlugin = createBlockDndPlugin();
+  const resizePlugin = createResizeDecoration({
+    horizontal: 'absolute',
+    minWidth: 350,
+  });
+  // const dndPlugin = createBlockDndPlugin();
 
-  const wixPluginsDecorators = composeDecorators(focusPlugin.decorator, dndPlugin.decorator);
+  const wixPluginsDecorators = composeDecorators(resizePlugin.decorator, focusPlugin.decorator);
+
   const wixPluginConfig = {
     decorator: wixPluginsDecorators,
     helpers,
@@ -53,7 +59,7 @@ const createPlugins = ({
     ];
   });
 
-  const pluginInstances = [focusPlugin, dndPlugin, ...wixPlugins];
+  const pluginInstances = [focusPlugin, resizePlugin, ...wixPlugins];
 
   return {
     pluginInstances,