+++ packages/plugin-gallery/web/src/components/gallery-controls/gallery-image-settings.js
@@ -34,8 +34,8 @@ class ImageSettings extends Component {
     this.props.onDeleteImage();
   }
 
-  replaceItem = event => {
-    this.props.handleFileChange(event);
+  replaceItem = files => {
+    this.props.handleFileChange(files);
   };
 
   getImageUrl = image =>