+++ packages/plugin-image/web/src/createImagePlugin.js
@@ -1,5 +1,9 @@
 import createToolbar from './toolbar';
-import { createBasePlugin } from 'wix-rich-content-common';
+import {
+  createBasePlugin,
+  PLUGIN_DECORATION_PROPS,
+  PLUGIN_DECORATIONS,
+} from 'wix-rich-content-common';
 import { Component } from './image-component';
 import { IMAGE_TYPE, IMAGE_TYPE_LEGACY } from './types';
 
@@ -20,6 +24,23 @@ const createImagePlugin = (config = {}) => {
     component: Component,
     type: IMAGE_TYPE,
     legacyType: IMAGE_TYPE_LEGACY,
+    pluginDecorationProps: (props, componentData) => {
+      const resizeableProps = PLUGIN_DECORATION_PROPS[PLUGIN_DECORATIONS.RESIZEABLE](props);
+      const { size } = componentData.config;
+      const isInlineSize = size === 'inline';
+      const { width, ...rest } = resizeableProps.style; // eslint-disable-line no-unused-vars
+      const style = isInlineSize ? resizeableProps.style : { ...rest };
+      return { ...resizeableProps, style };
+    },
+    componentWillReceiveDecorationProps: (props, nextProps, onPropsChange) => {
+      const { width } = PLUGIN_DECORATION_PROPS[PLUGIN_DECORATIONS.RESIZEABLE](props);
+      const { width: nextWidth } = PLUGIN_DECORATION_PROPS[PLUGIN_DECORATIONS.RESIZEABLE](
+        nextProps
+      );
+      if (width !== nextWidth) {
+        onPropsChange({ size: 'inline' });
+      }
+    },
     toolbar: createToolbar({
       helpers,
       anchorTarget,
@@ -27,6 +48,7 @@ const createImagePlugin = (config = {}) => {
       t,
       uiSettings,
       isMobile,
+      settings,
     }),
     helpers,
     anchorTarget,