+++ packages/viewer/web/src/utils/convertContentState.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { renderToStaticMarkup } from 'react-dom/server';
+import { convertFromRaw } from '@wix/draft-js';
 import { BLOCK_TYPES } from 'wix-rich-content-common';
 import redraft from 'redraft';
 import classNames from 'classnames';
@@ -115,6 +116,7 @@ const combineMappers = (mappers, ...args) => {
 };
 
 const redraftOptions = {
+  convertFromRaw,
   cleanup: {
     after: BLOCK_TYPES.filter(t => t.indexOf('header') === -1),
     split: true,